navbar on the bottom of all pages. slight changes to upload page and splash page layout. fixed logout button on navbar. added demo login button to login page
