+++ src/components/Nav.js
@@ -10,11 +10,10 @@ import IconButton from '@material-ui/core/IconButton';
 import { sendLogoutReq } from '../redux/user';
 import Grid from '@material-ui/core/Grid';
 
+
 const useStyles = makeStyles((theme) => ({
   root: {
-    flexGrow: 1,
     backgroundImage: theme.gradientBackground,
-    marginBottom: 64
   },
   menuButton: {
     marginRight: theme.spacing(2),
@@ -23,7 +22,7 @@ const useStyles = makeStyles((theme) => ({
     flexGrow: 1,
   },
   logout: {
-    marginLeft: 100,
+    width: 'min-content',
     // justifyItems: 'end',
     // alignItems: 'end'
   }
@@ -38,7 +37,7 @@ const NavBar = (props) => {
 
   const navigation = props.currentUserId ? (
 
-    <Grid container spacing={3}>
+    <Grid container spacing={3} style={{ 'justify-content': 'space-between' }}>
       <Grid item xs={10}>
         <NavLink style={{ color: 'white' }} to="/">
           <Button color="inherit">Instantelegram</Button>
@@ -50,7 +49,7 @@ const NavBar = (props) => {
           <Button color="inherit">Upload</Button>
         </NavLink>
       </Grid>
-      <Grid item xs={2}>
+      <Grid item>
         <div className={classes.logout} style={{ color: 'white' }}>
           <Button className={classes.logout} color="inherit" onClick={logOut}>Logout</Button>
         </div>
@@ -69,15 +68,13 @@ const NavBar = (props) => {
 
 
   return (
-    <div className={classes.root}>
-      <AppBar position="fixed" className={classes.root}>
-        <Toolbar>
-          <IconButton edge="start" className={classes.menuButton} color="inherit" aria-label="menu">
-          </IconButton>
-          {navigation}
-        </Toolbar>
-      </AppBar>
-    </div>
+    <AppBar position="fixed" className={classes.root}>
+      <Toolbar>
+        <IconButton edge="start" className={classes.menuButton} color="inherit" aria-label="menu">
+        </IconButton>
+        {navigation}
+      </Toolbar>
+    </AppBar>
   );
 }
