+++ src/config.js
@@ -1,9 +1,9 @@
 
+const production = process.env.NODE_ENV === "production";
+const apiBaseUrl = production ? "https://instantelegramback.herokuapp.com" : process.env.REACT_APP_API_BASE_URL;
+
 module.exports = {
-    apiBaseUrl: process.env.REACT_APP_API_BASE_URL,
+    apiBaseUrl: apiBaseUrl,
     cloudinaryUrl: process.env.REACT_APP_CLOUDINARY_URL,
     cloudinaryPreset: process.env.REACT_APP_CLOUDINARY_UPLOAD_PRESET,
 }
-// const production = process.env.NODE_ENV === "production";
-// export const apiBaseUrl = production ? "edit later" : process.env.REACT_APP_API_BASE_URL;
-
